requirements.txt
comtypes==1.4.6
numpy==2.1.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pandas==2.2.2
pillow==10.4.0
# pypiwin32
python-dateutil==2.9.0.post0
pyttsx3==2.91
pytz==2024.1
six==1.16.0
tzdata==2024.1
